csaファイルをGCSに保存するまでの実装完了
