+++ Pipfile
@@ -10,6 +10,7 @@ numpy = "*"
 black = "*"
 flake8 = "*"
 requests = "*"
+google-cloud-storage = "*"
 
 [dev-packages]
