+++ app/shogi.py
@@ -1,6 +1,7 @@
 import streamlit as st
 
 import utils
+from get_kif_file import CsaFetcher
 
 st.title("将棋アプリ解析君")
 
@@ -9,4 +10,6 @@ submit_button = st.button(label="評価値を計算する")
 
 if submit_button:
     csa_file_url = utils.check_input_text(text)
-    st.write(csa_file_url)
+    csa_fetcher = CsaFetcher(bucket="sasaki-sample")
+    csa_fetcher.fetch_csa_file(url=csa_file_url, upload_dir_path="")
+    st.write("upload done!")
