+++ dockerfile
@@ -1,0 +1,7 @@
+FROM python:3.9-slim
+COPY requirements.txt ./requirements.txt
+RUN pip install -r requirements.txt
+ENV APP_HOME /app
+WORKDIR $APP_HOME
+COPY . ./
+CMD streamlit run app/shogi.py --server.port 80
