+++ requirements.txt
@@ -1,0 +1,111 @@
+#
+# These requirements were autogenerated by pipenv
+# To regenerate from the project's Pipfile, run:
+#
+#    pipenv lock --requirements
+#
+
+-i https://pypi.org/simple
+altair==4.1.0; python_version >= '3.6'
+appnope==0.1.2; platform_system == 'Darwin'
+argon2-cffi-bindings==21.2.0; python_version >= '3.6'
+argon2-cffi==21.3.0; python_version >= '3.6'
+astor==0.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
+attrs==21.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
+backcall==0.2.0
+base58==2.1.1; python_version >= '3.5'
+black==21.12b0
+bleach==4.1.0; python_version >= '3.6'
+blinker==1.4
+cachetools==4.2.4; python_version ~= '3.5'
+certifi==2021.10.8
+cffi==1.15.0
+charset-normalizer==2.0.9; python_version >= '3'
+click==7.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
+debugpy==1.5.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
+decorator==5.1.0; python_version >= '3.5'
+defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
+entrypoints==0.3; python_version >= '2.7'
+flake8==4.0.1
+gitdb==4.0.9; python_version >= '3.6'
+gitpython==3.1.24; python_version >= '3.7'
+google-api-core==2.3.0; python_version >= '3.6'
+google-auth==2.3.3; python_version >= '3.6'
+google-cloud-core==2.2.1; python_version >= '3.6'
+google-cloud-storage==1.43.0
+google-crc32c==1.3.0; python_version >= '3.6'
+google-resumable-media==2.1.0; python_version >= '3.6'
+googleapis-common-protos==1.54.0; python_version >= '3.6'
+idna==3.3; python_version >= '3'
+ipykernel==6.6.0; python_version >= '3.4'
+ipython-genutils==0.2.0
+ipython==7.30.1; python_version >= '3.7'
+ipywidgets==7.6.5
+jedi==0.18.1; python_version >= '3.6'
+jinja2==3.0.3; python_version >= '3.6'
+jsonschema==4.2.1; python_version >= '3.7'
+jupyter-client==7.1.0; python_full_version >= '3.6.1'
+jupyter-core==4.9.1; python_version >= '3.6'
+jupyterlab-pygments==0.1.2
+jupyterlab-widgets==1.0.2; python_version >= '3.6'
+markupsafe==2.0.1; python_version >= '3.6'
+matplotlib-inline==0.1.3; python_version >= '3.5'
+mccabe==0.6.1
+mistune==0.8.4
+mypy-extensions==0.4.3
+nbclient==0.5.9; python_full_version >= '3.6.1'
+nbconvert==6.3.0; python_version >= '3.7'
+nbformat==5.1.3; python_version >= '3.5'
+nest-asyncio==1.5.4; python_version >= '3.5'
+notebook==6.4.6; python_version >= '3.6'
+numpy==1.21.4
+packaging==21.3; python_version >= '3.6'
+pandas==1.3.5
+pandocfilters==1.5.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
+parso==0.8.3; python_version >= '3.6'
+pathspec==0.9.0
+pexpect==4.8.0; sys_platform != 'win32'
+pickleshare==0.7.5
+pillow==8.4.0; python_version >= '3.6'
+platformdirs==2.4.0; python_version >= '3.6'
+prometheus-client==0.12.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
+prompt-toolkit==3.0.24; python_full_version >= '3.6.2'
+protobuf==3.19.1; python_version >= '3.5'
+ptyprocess==0.7.0
+pyarrow==6.0.1; python_version >= '3.6'
+pyasn1-modules==0.2.8
+pyasn1==0.4.8
+pycodestyle==2.8.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
+pycparser==2.21
+pydeck==0.7.1; python_version >= '3.7'
+pyflakes==2.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
+pygments==2.10.0; python_version >= '3.5'
+pympler==0.9
+pyparsing==3.0.6; python_version >= '3.6'
+pyrsistent==0.18.0; python_version >= '3.6'
+python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
+pytz-deprecation-shim==0.1.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
+pytz==2021.3
+pyzmq==22.3.0; python_version >= '3.6'
+requests==2.26.0
+rsa==4.8; python_version >= '3.6'
+send2trash==1.8.0
+setuptools==59.6.0; python_version >= '3.6'
+six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
+smmap==5.0.0; python_version >= '3.6'
+streamlit==1.2.0
+terminado==0.12.1; python_version >= '3.6'
+testpath==0.5.0; python_version >= '3.5'
+toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
+tomli==1.2.2; python_version >= '3.6'
+toolz==0.11.2; python_version >= '3.5'
+tornado==6.1; python_version >= '3.5'
+traitlets==5.1.1; python_version >= '3.7'
+typing-extensions==4.0.1; python_version >= '3.6'
+tzdata==2021.5; python_version >= '3.6'
+tzlocal==4.1; python_version >= '3.6'
+urllib3==1.26.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
+validators==0.18.2; python_version >= '3.4'
+wcwidth==0.2.5
+webencodings==0.5.1
+widgetsnbextension==3.5.2
